+++ vpc-20160428/src/Models/DescribeExpressConnectTrafficQosRequest.php
@@ -4,6 +4,7 @@
 
 namespace AlibabaCloud\SDK\Vpc\V20160428\Models;
 
+use AlibabaCloud\SDK\Vpc\V20160428\Models\DescribeExpressConnectTrafficQosRequest\tags;
 use AlibabaCloud\Tea\Model;
 
 class DescribeExpressConnectTrafficQosRequest extends Model
@@ -61,7 +62,17 @@ class DescribeExpressConnectTrafficQosRequest extends Model
     /**
      * @var string
      */
+    public $resourceGroupId;
+
+    /**
+     * @var string
+     */
     public $resourceOwnerAccount;
+
+    /**
+     * @var tags[]
+     */
+    public $tags;
     protected $_name = [
         'clientToken'          => 'ClientToken',
         'maxResults'           => 'MaxResults',
@@ -71,7 +82,9 @@ class DescribeExpressConnectTrafficQosRequest extends Model
         'qosIdList'            => 'QosIdList',
         'qosNameList'          => 'QosNameList',
         'regionId'             => 'RegionId',
+        'resourceGroupId'      => 'ResourceGroupId',
         'resourceOwnerAccount' => 'ResourceOwnerAccount',
+        'tags'                 => 'Tags',
     ];
 
     public function validate()
@@ -105,8 +118,20 @@ class DescribeExpressConnectTrafficQosRequest extends Model
         if (null !== $this->regionId) {
             $res['RegionId'] = $this->regionId;
         }
+        if (null !== $this->resourceGroupId) {
+            $res['ResourceGroupId'] = $this->resourceGroupId;
+        }
         if (null !== $this->resourceOwnerAccount) {
             $res['ResourceOwnerAccount'] = $this->resourceOwnerAccount;
+        }
+        if (null !== $this->tags) {
+            $res['Tags'] = [];
+            if (null !== $this->tags && \is_array($this->tags)) {
+                $n = 0;
+                foreach ($this->tags as $item) {
+                    $res['Tags'][$n++] = null !== $item ? $item->toMap() : $item;
+                }
+            }
         }
 
         return $res;
@@ -148,8 +173,20 @@ class DescribeExpressConnectTrafficQosRequest extends Model
         if (isset($map['RegionId'])) {
             $model->regionId = $map['RegionId'];
         }
+        if (isset($map['ResourceGroupId'])) {
+            $model->resourceGroupId = $map['ResourceGroupId'];
+        }
         if (isset($map['ResourceOwnerAccount'])) {
             $model->resourceOwnerAccount = $map['ResourceOwnerAccount'];
+        }
+        if (isset($map['Tags'])) {
+            if (!empty($map['Tags'])) {
+                $model->tags = [];
+                $n           = 0;
+                foreach ($map['Tags'] as $item) {
+                    $model->tags[$n++] = null !== $item ? tags::fromMap($item) : $item;
+                }
+            }
         }
 
         return $model;