+++ vpc-20160428/src/Models/DescribeExpressConnectTrafficQosResponseBody/qosList.php
@@ -6,6 +6,7 @@ namespace AlibabaCloud\SDK\Vpc\V20160428\Models\DescribeExpressConnectTrafficQos
 
 use AlibabaCloud\SDK\Vpc\V20160428\Models\DescribeExpressConnectTrafficQosResponseBody\qosList\associatedInstanceList;
 use AlibabaCloud\SDK\Vpc\V20160428\Models\DescribeExpressConnectTrafficQosResponseBody\qosList\queueList;
+use AlibabaCloud\SDK\Vpc\V20160428\Models\DescribeExpressConnectTrafficQosResponseBody\qosList\tags;
 use AlibabaCloud\Tea\Model;
 
 class qosList extends Model
@@ -49,11 +50,21 @@ class qosList extends Model
     public $queueList;
 
     /**
+     * @var string
+     */
+    public $resourceGroupId;
+
+    /**
      * @example Normal
      *
      * @var string
      */
     public $status;
+
+    /**
+     * @var tags[]
+     */
+    public $tags;
     protected $_name = [
         'associatedInstanceList' => 'AssociatedInstanceList',
         'progressing'            => 'Progressing',
@@ -61,7 +72,9 @@ class qosList extends Model
         'qosId'                  => 'QosId',
         'qosName'                => 'QosName',
         'queueList'              => 'QueueList',
+        'resourceGroupId'        => 'ResourceGroupId',
         'status'                 => 'Status',
+        'tags'                   => 'Tags',
     ];
 
     public function validate()
@@ -101,8 +114,20 @@ class qosList extends Model
                 }
             }
         }
+        if (null !== $this->resourceGroupId) {
+            $res['ResourceGroupId'] = $this->resourceGroupId;
+        }
         if (null !== $this->status) {
             $res['Status'] = $this->status;
+        }
+        if (null !== $this->tags) {
+            $res['Tags'] = [];
+            if (null !== $this->tags && \is_array($this->tags)) {
+                $n = 0;
+                foreach ($this->tags as $item) {
+                    $res['Tags'][$n++] = null !== $item ? $item->toMap() : $item;
+                }
+            }
         }
 
         return $res;
@@ -146,8 +171,20 @@ class qosList extends Model
                 }
             }
         }
+        if (isset($map['ResourceGroupId'])) {
+            $model->resourceGroupId = $map['ResourceGroupId'];
+        }
         if (isset($map['Status'])) {
             $model->status = $map['Status'];
+        }
+        if (isset($map['Tags'])) {
+            if (!empty($map['Tags'])) {
+                $model->tags = [];
+                $n           = 0;
+                foreach ($map['Tags'] as $item) {
+                    $model->tags[$n++] = null !== $item ? tags::fromMap($item) : $item;
+                }
+            }
         }
 
         return $model;